requirements.txt
python-dotenv==1.0.0
python-telegram-bot==13.1
loguru==0.7.0
notifiers==1.3.3